requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioresponses==0.7.6
aiosignal==1.3.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
awscli==1.32.39
backoff==2.2.1
blinker==1.7.0
boto3==1.34.39
botocore==1.34.39
CacheControl==0.14.0
cachetools==5.3.2
catboost==1.2.3
chardet==3.0.4
click==8.1.3
colorama==0.4.4
comtypes==1.2.0
contourpy==1.1.1
cuda-python==12.3.0
cvzone==1.6.1
cycler==0.12.0
Cython==3.0.7
dataclasses-json==0.6.4
deepface==0.0.84
Django==4.2.8
django-crispy-forms==2.1
docker==7.0.0
docopt==0.6.2
docutils==0.16
easyocr==1.7.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
fire==0.5.0
firebase-admin==6.4.0
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.4.1
fsspec==2023.12.2
future==1.0.0
gast==0.5.4
gdown==5.1.0
geocoder==1.38.1
glfw==2.5.9
googletrans==4.0.0rc1
gspread==6.0.1
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.8
imageio==2.33.1
importlib-resources==6.1.1
imutils==0.5.4
inference-sdk==0.9.16
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
Js2Py==0.74
keras==2.15.0
keyring==8.7
keyrings.alt==5.0.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
libretranslatepy==2.1.1
lockfile==0.12.2
lxml==5.1.0
Markdown==3.5.2
marshmallow==3.21.1
matplotlib==3.8.0
mediapipe==0.10.5
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
mtcnn==0.1.1
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==9.4.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pygame==2.5.2
pyglet==2.0.7
pyjsparser==2.7.1
PyJWT==2.8.0
pyparsing==3.1.1
pypiwin32==223
PyPrind==2.11.3
pyserial==3.5
pySerialTransfer==2.6.9
pySmartDL==1.3.4
pytesseract==0.3.10
python-bidi==0.4.2
pyttsx3==2.90
pyttsx4==3.0.15
pywin32==306
pywin32-ctypes==0.2.2
ratelim==0.1.6
regex==2023.12.25
reportlab==4.1.0
requests-oauthlib==1.3.1
retina-face==0.0.14
rfc3986==1.5.0
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.13.2
shapely==2.0.3
simplegeneric==0.8.1
sounddevice==0.4.6
SpeechRecognition==3.10.1
sqlparse==0.4.4
StrEnum==0.4.15
supervision==0.18.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
translate==3.6.1
twilio==9.0.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.1.13
uritemplate==4.1.1
voluptuous==0.14.2
webencodings==0.5.1
Werkzeug==3.0.1
word2number==1.1
wrapt==1.14.1
WTForms==3.0.1
yarl==1.9.4
yolo==0.3.1

